[refactor]필요 없는 파일들 삭제
